RLHF/golang_alignment_5/23-10-24/105378/turn1/modelB/main.go
// Step 1: Define the DSL structures and functions

package compressiondsl

// CompressionScheme represents a compression scheme
type CompressionScheme struct {
	Name string
	// Add other fields as needed for your DSL
}

// Compressor represents a function that compresses data
type Compressor func([]byte) []byte

// Define the DSL functions
func CompressWith(scheme CompressionScheme) Compressor {
	// Implement the compression logic based on the scheme
	// For simplicity, we'll just return an identity function in this example
	return func(data []byte) []byte {
		return data
	}
}
