RLHF/golang_alignment_5/23-10-24/105378/turn1/modelB/main_test.go
// Step 2: Write the tests for the type inference functionality

package compressiondsl

import (
	"testing"
)

func TestInferDataType(t *testing.T) {
	// Define some test cases
	testCases := []struct {
		name     string
		scheme   CompressionScheme
		data     interface{}
		expected string
	}{
		{
			name:     "Infer string data",
			scheme:   CompressionScheme{Name: "LZ4"},
			data:     "Hello, World!",
			expected: "string",
		},
		{
			name:     "Infer integer data",
			scheme:   CompressionScheme{Name: "Deflate"},
			data:     42,
			expected: "int",
		},
		// Add more test cases for different data types and compression schemes
	}

	for _, tc := range testCases {
		t.Run(tc.name, func(t *testing.T) {
			// Call the function under test (not yet implemented)
			actual := inferDataType(tc.scheme, tc.data)
			if actual != tc.expected {
				t.Errorf("Expected data type %q, got %q", tc.expected, actual)
			}
		})
	}
}

// Step 3: Implement the type inference function

func inferDataType(scheme CompressionScheme, data interface{}) string {
	// Implement the type inference logic based on the data type and compression scheme
	// For this example, we'll just perform basic type checking
	switch data.(type) {
	case string:
		return "string"
	case int:
		return "int"
	default:
		return "unknown"
	}
}
